Adding notificationMixin
